Moved generateReports into core.database.py
