+++ backend/app/api/v1/endpoints/student.py
@@ -3,8 +3,8 @@ from sqlalchemy.orm import Session
 import json
 import numpy as np
 from typing import Dict, List, Tuple
-from ....schemas.pydantic_base_models.user_schemas import StudentSchema
-from ....core.database import get_db, link_logininfo
+from ....schemas.reportschema import StudentReport
+from ....core.database import get_db, link_logininfo, generateStudentInformationReport
 from app.models import Student, LoginInfo as User
 from app.core.security import (
     get_current_active_user
@@ -51,4 +51,42 @@ async def update_login(
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
             detail="An error occurred while processing your request"
         )
+
+@router.get("/info", response_model=StudentReport)
+async def generate_report(
+    current_user: User = Depends(get_current_active_user),
+    db: Session = Depends(get_db)
+):
+    try:
+        print(current_user.username)
+        if current_user.issuperuser:
+            raise HTTPException(
+                status_code=403,
+                detail="Admins accounts can not access student reports route"
+            )
+
+        studentid = db.query(Student.studentid).filter(Student.logininfoid == current_user.logininfoid).scalar()
+        if not studentid:
+            raise HTTPException(
+                status_code=404,
+                detail="Login Info is not Attached to a Student"
+            )
+        
+        studentinfo = generateStudentInformationReport(studentid, db)
+        if not studentinfo:
+            raise HTTPException(
+                status_code=404,
+                detail="No Student Info Found"
+            )
+        
+        return studentinfo
     
+    except HTTPException:
+        raise
+    
+    except Exception as e:
+        print(f"Login error: {str(e)}")
+        raise HTTPException(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            detail="An error occurred while processing your request"
+        )
