+++ backend/app/core/database.py
@@ -4,6 +4,16 @@ from app.models import Student, LoginInfo
 from .config import settings
 from .base import Base
 import math
+from app.models import (
+    Student,
+    GraduationStatus,
+    Exam,
+    ExamResults,
+    GradeClassification,
+    StudentGrade,
+    ClassOffering
+)
+from app.schemas.reportschema import StudentReport, ExamReport, GradeReport, StudentCompleteReport
 
 engine = create_engine(settings.sync_database_url)
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
@@ -97,4 +107,96 @@ def link_logininfo(studentid, logininfoid):
         
     db.close()
     
+def generateStudentInformationReport(student_id, db):
+    data = db.query(
+        Student.studentid,
+        Student.lastname,
+        Student.firstname,
+        Student.cumgpa,
+        Student.bcpmgpa,
+        Student.mmicalc,
+        GraduationStatus.rosteryear,
+        GraduationStatus.graduationyear,
+        GraduationStatus.graduated,
+        GraduationStatus.graduationlength,
+        GraduationStatus.status
+    ).join(
+        GraduationStatus, Student.studentid == GraduationStatus.studentid
+    ).filter(
+        Student.studentid == student_id
+    ).first()
     
+    '''
+    Pydantic handeling Float(NaN) to None conversion will not need to handle
+    multiple rows as a student will only have a singular student report
+    '''
+    studentinformation = StudentReport(
+        StudentID = data.studentid,
+        LastName = data.lastname,
+        FirstName = data.firstname,
+        CumGPA = data.bcpmgpa,
+        MMICalc = data.mmicalc,
+        RosterYear = data.rosteryear,
+        GraduationYear = data.graduationyear,
+        Graduated = data.graduated,
+        GraduationLength = data.graduationlength,
+        Status = data.status
+    )
+    
+    return studentinformation  
+
+def generateExamReport(student_id, db):
+    data = db.query(
+        Exam.examname,
+        ExamResults.score,
+        Exam.passscore,
+        ExamResults.passorfail
+    ).join(
+        Exam, ExamResults.examid == Exam.examid
+    ).filter(
+        ExamResults.studentid == student_id
+    ).all()
+    
+    exams = []
+    #Converting to a dictionary as there can be multiple records of exams
+    for row in data:
+        exam_dict = {
+            "ExamName": row[0],
+            "Score": row[1],
+            "PassScore": row[2],
+            "PassOrFail": row[3]
+        }
+        pydanticData = ExamReport(**exam_dict)
+        exams.append(pydanticData)
+        
+    return exams
+
+def generateGradeReport(student_id, db):
+    data = db.query(
+        GradeClassification.classificationname,
+        StudentGrade.pointsearned,
+        StudentGrade.pointsavailable,
+        ClassOffering.classid,
+        ClassOffering.datetaught
+    ).join(
+        GradeClassification, StudentGrade.gradeclassificationid == GradeClassification.gradeclassificationid
+    ).join(
+        ClassOffering, GradeClassification.classofferingid == ClassOffering.classofferingid
+    ).filter(
+        StudentGrade.studentid == student_id
+    ).all()
+    
+    grades = []
+    #Converting to dictionary as there can be multiple records of grades
+    for row in data:
+        grade_dict = {
+            "ClassificationName": row[0],
+            "PointsEarned": row[1],
+            "PointsAvailable": row[2],
+            "ClassID": row[3],
+            "DateTaught": row[4]
+        }
+        pydanticData = GradeReport(**grade_dict)
+        grades.append(pydanticData)
+        
+    return grades
